tests for inserting into collections
